[kcalcore] Fallback to UTC timespec if datetime validation fails during transition generation. Contributes to MER#1011
Some CalDAV servers provide VTIMEZONE definitions for CEST which have
DAYLIGHT component which includes a YEARLY repeating RRULE from a
particular DTSTART.  This works fine if the DTSTART value is in or
after 1981, but otherwise will fail (presumably due to differing
system definition for the timespec).

The result is that dates created from the RRULE using a timespec
built from the VTIMEZONE definition (ie, the DST transition dates)
will be deemed invalid by KDateTime.isValid().  The result of this
is that they will not be added to the list of transitions.

This commit ensures that if such a circumstance is encountered, the
transition datetime is built using UTC timespec instead if possible
(ie, if the utc offset of the custom timespec is zero anyway).
This datetime is then deemed valid, and able to be used as a DST
transition datetime.

Contributes to MER#1011